Itemのバリデーション

+++ db/migrate/20200126053317_create_items.rb
@@ -3,7 +3,7 @@ class CreateItems < ActiveRecord::Migration[5.2]
     create_table :items do |t|
       t.string     :title,    null: false
       t.text       :details,  null: false
-      t.integer    :price,    null: false
+      t.bigint     :price,    null: false
       t.integer    :seller_id
       t.references :user, foreign_key:
